requirements.txt
pandas>=2.2.0
pyyaml>=6.0.0
huggingface_hub>=0.31.0
